Improve the nvidia GPU check.
Try to examine the output of "nvidia-smi --list-gpus" to see if any
GPUs were found. If any were found, but no CDI configuration file is
found, print a message with a pointer to documentation of how to
generate a CDI configuration file.

Add an optional encoding argument to run_cmd() to facilitate checking
the command output.

Signed-off-by: John Wiele <jwiele@redhat.com>